requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Database & ORM
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# HTTP Client & Downloads
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0
urllib3==2.1.0

# Media Processing & M3U8
m3u8==4.0.0
ffmpeg-python==0.2.0

# Task Queue & Background Jobs
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WebSockets for real-time updates
channels==4.0.0
channels-redis==4.1.0

# File Management
django-storages==1.14.2
boto3==1.34.0

# API Documentation
drf-spectacular==0.26.5

# Internationalization
django-modeltranslation==0.18.11

# Configuration & Environment
python-dotenv==1.0.0
django-environ==0.11.2

# Development & Testing
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Monitoring & Logging
sentry-sdk==1.38.0

# Utils
python-slugify==8.0.1
pillow==10.1.0